refac(app): Move deviceGetter interface to separate file
Changelog: None
